backend/requirements.txt
Flask==2.2.3
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
torch==2.0.0+cu118
